requirements/base.txt
# Base requirements for all components
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
aiofiles>=23.0.0
python-multipart>=0.0.6# M
onitoring dependencies
prometheus-client==0.19.0
flask==3.0.0
psutil==5.9.6